requirements.txt
# Lists all the python requirements for the project
# Django>=3.2.4,<3.3 #Specifies to use latest patch version i.e the .4 but will prevent from using the next minor version ie .3
# djangorestframework>=3.12.4,3.13 # Specicfies to use the latest patch version of 3.12 but not the next minor versoion ie 3.13.
Django>=4.0.4,<4.1
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.22.1,<0.23 # https://drf-spectacular.readthedocs.io/en/latest/index.html